Replace empty cells with zero
This function currently fails if `fprintf` receives an empty cell. This is an edit I've made in my copy and it seems to work fine. If you would like an example file for when palm fails, I can upload those so that you can see the error. 

Of note, this does not appear to be a problem for palm116 in octave as the same command runs for palm116 and octave that fails for palm119 and matlab 2020